refactor(container): mv ArticleXXX/Labeler to unit scope
