+++ src/containers/editor/PostEditor/EditorFooter.js
@@ -4,7 +4,7 @@ import { pluck } from 'ramda'
 import { ICON_CMD } from '@/config'
 
 import DocUploader from '@/containers/tool/DocUploader'
-import Labeler from '@/containers/Labeler'
+import Labeler from '@/containers/unit/Labeler'
 import Maybe from '@/components/Maybe'
 
 import {
